refactored reducers
