+++ package.json
@@ -61,12 +61,12 @@
     "webpack-dev-server": "^1.16.2"
   },
   "dependencies": {
+    "flow-bin": "^0.33.0",
     "jsoneditor": "^5.5.10",
     "material-ui": "^0.16.4",
-    "react-tap-event-plugin": "^2.0.1",
-    "sprintf-js": "^1.0.3",
-    "flow-bin": "^0.33.0",
+    "normalizr": "^3.1.0",
     "react": "^15.4.0",
+    "react-chartjs-2": "1.5.1",
     "react-dom": "^15.4.0",
     "react-hot-loader": "^3.0.0-beta.5",
     "react-json": "^0.2.1",
@@ -74,12 +74,13 @@
     "react-json-pretty": "^1.2.1",
     "react-redux": "^4.4.5",
     "react-router": "^2.8.1",
+    "react-tap-event-plugin": "^2.0.1",
     "react-tools": "^0.13.3",
     "redux": "^3.6.0",
     "redux-devtools": "^3.3.1",
     "redux-promise-middleware": "^4.1.0",
     "redux-thunk": "^2.1.0",
-    "whatwg-fetch": "^1.0.0",
-    "react-chartjs-2": "1.5.1"
+    "sprintf-js": "^1.0.3",
+    "whatwg-fetch": "^1.0.0"
   }
 }
